+++ client/src/components/Typography.js
@@ -5,6 +5,11 @@ export const Header = styled.h1`
 
 export const PlainText = styled.p`
 `;
+
 export const BoldText = styled.p`
   font-weight: 700;
 `;
+
+export const Title = styled.h4`
+  margin-top: 20px;
+`