+++ server/server.js
@@ -1,18 +1,36 @@
 const express = require('express');
-const bodyParser = require('body-parser')
+const bodyParser = require('body-parser');
+
+var Sequelize = require('sequelize');
 
 const app = express();
 const port = process.env.PORT || 3001;
 
 app.use(bodyParser());
 
+const sequelize = new Sequelize('flom-dev', 'flom', 'flom', {
+  dialect: 'postgres',
+});
+
 app.get('/api/hello', (req, res) => {
-  res.send({ express: 'Hello From Express' });
+  res.send({ express: 'Message from Express server' });
 });
 
 app.post('/api/echo', (req, res) => {
   console.log('Recieved echo request with body:', req.body);
-  res.end({ express: req.body.text });
+  res.json(req.body);
+});
+
+app.post('/api/_unsafe_sqlTest', (req, res) => {
+  console.log('Recieved sql test request with body:', req.body);
+  sequelize.query(req.body.query).spread((results, metadata) => {
+    console.log('results', results);
+    console.log('metadata', metadata);
+    res.json({
+      results: results,
+      metadata: metadata
+    })
+  })
 });
 
 if (process.env.NODE_ENV === 'production') {